Add files via upload
It is a simple game to play, but a bit complex in its creation. It was built with HTML, JS, CSS, JSON and NodeJS. In this game you can interact with another player from another device (as long as you are both connected to the same Wi-Fi network).
You will have to choose one of the characters and your friend will do it too. Both characters appear on a small map where, after colliding against each other, they will challenge each other to a power duel.
Afterwards, each of you will have to choose the attacks with which your "pet" will fight. And... may the best one win!

Instructions for playing:
1. Download the files to your PC
2. You will need to install NodeJS on your computer. You can search the net for information on how to do it. It is easy.
3. Start the NodeJS application. On the same page where you learned how to install the package, it will also tell you how to start it.
4. In the configuration menu of your Wi-fi network, you can look up your IPv4 address, then, in your browser, copy that address and add ":8080" at the end.
("For example: 192.168.13.67:8080". This will give you access to the game.
Remember that your partner will have to do the same from his device, and it will have to be connected to the same Wi-fi network).
5. Play and enjoy

Best regards!